Rename package(core_navigation to core.navigation)
