+++ core-navigation/src/main/java/com/hanbitkang/core/navigation/MpNavigationDestination.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.core_navigation
+package com.hanbitkang.core.navigation
 
 interface MpNavigationDestination {
     val route: String
